feat: implementar 8 network security checks e endpoint consolidado
- Dividir SSH probe em dois probes especializados:
  - ssh_weak_cipher: detecta cifras fracas (CBC, 3DES, arcfour)
  - ssh_weak_mac: detecta MACs fracos (MD5, SHA1-96, RIPEMD)
- Adicionar endpoint GET /api/v1/scans/{scan_id}/network
- Endpoint retorna status consolidado dos 8 checks de segurança
- Cada check retorna status "ok" ou "risk" com evidence
- Adicionar contract tests para validar API
- Corrigir bug de bit shifting no RDP probe
- Atualizar documentação completa (README, HOWTO_DEV, Swagger)
- Adicionar exemplos em curl_examples.sh

Closes #13

+++ internal/probes/rdp.go
@@ -104,7 +104,7 @@ func (p *RDPProbe) Probe(ctx context.Context, ip string, port int) (*models.Prob
 	for i := 7; i < n-4; i++ {
 		if response[i] == 0x02 { // TYPE_RDP_NEG_RSP
 			if i+7 < n {
-				selectedProtocol := response[i+4] | (response[i+5] << 8) | (response[i+6] << 16) | (response[i+7] << 24)
+				selectedProtocol := uint32(response[i+4]) | (uint32(response[i+5]) << 8) | (uint32(response[i+6]) << 16) | (uint32(response[i+7]) << 24)
 				switch selectedProtocol {
 				case 0x00000000:
 					protocolDetected = "Standard RDP"
